feature-WebApp/backend/requirements.txt
flask==3.1.0
black==25.1.0
flake8==7.1.2
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.23
mysql-connector-python==8.0.33
python-dotenv==1.0.0
bcrypt==4.0.1
